requirements.txt
InquirerPy
python-dotenv